+++ src/Model/DataTablesGetDataPostRequest/WhereCondition.php
@@ -4,74 +4,27 @@ declare(strict_types=1);
 namespace Citilink\ExpertSenderApi\Model\DataTablesGetDataPostRequest;
 
 use Citilink\ExpertSenderApi\Enum\DataTablesGetDataPostRequest\Operator;
-use Webmozart\Assert\Assert;
 
 /**
- * Where condition for get table data
+ * Where condition
+ *
+ * @deprecated Use {@see \Citilink\ExpertSenderApi\Model\WhereCondition} instead
  *
  * @author Nikita Sapogov <sapogov.n@citilink.ru>
  */
-class WhereCondition
+class WhereCondition extends \Citilink\ExpertSenderApi\Model\WhereCondition
 {
     /**
-     * @var string Column name
-     */
-    private $columnName;
-
-    /**
-     * @var Operator Operator
-     */
-    private $operator;
-
-    /**
-     * @var string Value
-     */
-    private $value;
-
-    /**
-     * Constructor
+     * Constructor.
      *
      * @param string $columnName Column name
      * @param Operator $operator Operator
-     * @param string|int|float $value Value
+     * @param float|int|string $value Value
      */
-    public function __construct(string $columnName, Operator $operator, $value)
+    public function __construct($columnName, Operator $operator, $value)
     {
-        Assert::scalar($value);
-        Assert::notEmpty($columnName);
-        Assert::notEmpty($value);
-        $this->columnName = $columnName;
-        $this->operator = $operator;
-        $this->value = strval($value);
-    }
+        @trigger_error('use \Citilink\ExpertSenderApi\Model\WhereCondition instead', E_USER_DEPRECATED);
 
-    /**
-     * Get column name
-     *
-     * @return string Column name
-     */
-    public function getColumnName(): string
-    {
-        return $this->columnName;
-    }
-
-    /**
-     * Get operator
-     *
-     * @return Operator Operator
-     */
-    public function getOperator(): Operator
-    {
-        return $this->operator;
-    }
-
-    /**
-     * Get value
-     *
-     * @return string Value
-     */
-    public function getValue(): string
-    {
-        return $this->value;
+        parent::__construct($columnName, $operator, $value);
     }
 }